Change nhanes_data columns to DOUBLE PRECISION

+++ docker/postgres/init/01-init-nhanes-and-imdb.sql
@@ -6,14 +6,14 @@ CREATE DATABASE imdb;
 DROP TABLE IF EXISTS nhanes_data;
 
 CREATE TABLE nhanes_data (
-    seqn INTEGER,
+    seqn DOUBLE PRECISION,
     age_group TEXT,
-    ridageyr INTEGER,
-    riagendr INTEGER,
-    paq605 INTEGER,
+    ridageyr DOUBLE PRECISION,
+    riagendr DOUBLE PRECISION,
+    paq605 DOUBLE PRECISION,
     bmxbmi DOUBLE PRECISION,
     lbxglu DOUBLE PRECISION,
-    diq010 INTEGER,
+    diq010 DOUBLE PRECISION,
     lbxglt DOUBLE PRECISION,
     lbxin DOUBLE PRECISION
 );
